Renamed existing List* plugins as LegacyList* plugins.
