+++ packages/ckeditor5-list/src/legacylistproperties/legacylistreversedcommand.ts
@@ -4,21 +4,22 @@
  */
 
 /**
- * @module list/listproperties/liststartcommand
+ * @module list/legacylistproperties/legacylistreversedcommand
  */
 
 import { Command } from 'ckeditor5/src/core.js';
-import { getSelectedListItems } from '../list/utils.js';
+import { getSelectedListItems } from '../legacylist/legacyutils.js';
 
 /**
- * The list start index command. It changes the `listStart` attribute of the selected list items.
- * It is used by the {@link module:list/listproperties~ListProperties list properties feature}.
+ * The reversed list command. It changes the `listReversed` attribute of the selected list items. As a result, the list order will be
+ * reversed.
+ * It is used by the {@link module:list/legacylistproperties~LegacyListProperties legacy list properties feature}.
  */
-export default class ListStartCommand extends Command {
+export default class LegacyListReversedCommand extends Command {
 	/**
 	 * @inheritDoc
 	 */
-	declare public value: number | null;
+	declare public value: boolean | null;
 
 	/**
 	 * @inheritDoc
@@ -33,16 +34,16 @@ export default class ListStartCommand extends Command {
 	 * Executes the command.
 	 *
 	 * @fires execute
-	 * @param options.startIndex The list start index.
+	 * @param options.reversed Whether the list should be reversed.
 	 */
-	public override execute( { startIndex = 1 }: { startIndex?: number } = {} ): void {
+	public override execute( options: { reversed?: boolean } = {} ): void {
 		const model = this.editor.model;
 		const listItems = getSelectedListItems( model )
 			.filter( item => item.getAttribute( 'listType' ) == 'numbered' );
 
 		model.change( writer => {
 			for ( const item of listItems ) {
-				writer.setAttribute( 'listStart', startIndex >= 0 ? startIndex : 1, item );
+				writer.setAttribute( 'listReversed', !!options.reversed, item );
 			}
 		} );
 	}
@@ -56,7 +57,7 @@ export default class ListStartCommand extends Command {
 		const listItem = this.editor.model.document.selection.getFirstPosition()!.parent;
 
 		if ( listItem && listItem.is( 'element', 'listItem' ) && listItem.getAttribute( 'listType' ) == 'numbered' ) {
-			return listItem.getAttribute( 'listStart' ) as number;
+			return listItem.getAttribute( 'listReversed' ) as boolean;
 		}
 
 		return null;
