+++ packages/ckeditor5-list/src/legacylistproperties/legacyliststartcommand.ts
@@ -4,22 +4,21 @@
  */
 
 /**
- * @module list/listproperties/listreversedcommand
+ * @module list/legacylistproperties/legacyliststartcommand
  */
 
 import { Command } from 'ckeditor5/src/core.js';
-import { getSelectedListItems } from '../list/utils.js';
+import { getSelectedListItems } from '../legacylist/legacyutils.js';
 
 /**
- * The reversed list command. It changes the `listReversed` attribute of the selected list items. As a result, the list order will be
- * reversed.
- * It is used by the {@link module:list/listproperties~ListProperties list properties feature}.
+ * The list start index command. It changes the `listStart` attribute of the selected list items.
+ * It is used by the {@link module:list/legacylistproperties~LegacyListProperties legacy list properties feature}.
  */
-export default class ListReversedCommand extends Command {
+export default class LegacyListStartCommand extends Command {
 	/**
 	 * @inheritDoc
 	 */
-	declare public value: boolean | null;
+	declare public value: number | null;
 
 	/**
 	 * @inheritDoc
@@ -34,16 +33,16 @@ export default class ListReversedCommand extends Command {
 	 * Executes the command.
 	 *
 	 * @fires execute
-	 * @param options.reversed Whether the list should be reversed.
+	 * @param options.startIndex The list start index.
 	 */
-	public override execute( options: { reversed?: boolean } = {} ): void {
+	public override execute( { startIndex = 1 }: { startIndex?: number } = {} ): void {
 		const model = this.editor.model;
 		const listItems = getSelectedListItems( model )
 			.filter( item => item.getAttribute( 'listType' ) == 'numbered' );
 
 		model.change( writer => {
 			for ( const item of listItems ) {
-				writer.setAttribute( 'listReversed', !!options.reversed, item );
+				writer.setAttribute( 'listStart', startIndex >= 0 ? startIndex : 1, item );
 			}
 		} );
 	}
@@ -57,7 +56,7 @@ export default class ListReversedCommand extends Command {
 		const listItem = this.editor.model.document.selection.getFirstPosition()!.parent;
 
 		if ( listItem && listItem.is( 'element', 'listItem' ) && listItem.getAttribute( 'listType' ) == 'numbered' ) {
-			return listItem.getAttribute( 'listReversed' ) as boolean;
+			return listItem.getAttribute( 'listStart' ) as number;
 		}
 
 		return null;
